+++ changelog.md
@@ -1,2 +1,3 @@
-[0.0.2] : Fix issues regarding switch between 2d and 3d graph
 [0.0.1] : First version
+[0.0.2] : Fix issues regarding switch between 2d and 3d graph
+[0.0.3] : Fix issue regarding the creation of the cmake query even if the project have no CMakeLists.txt